+++ src/main/java/com/example/ecopr/service/ProductService.java
@@ -4,5 +4,5 @@ import com.example.ecopr.model.Products;
 import java.util.List;
 
 public interface ProductService {
-    List<Products> getProductsByCategory(int categoryId); // Changed to int
+    List<Products> getProductsByCategory(Long categoryId);
 }
